Add files via upload
This was a project for a class about the history of games. The project was to create a study guide for the class of a certain chapter from the textbook. Most of the other students opted to make Prezis or PowerPoints. The inspiration for this was since the first type of games on the computer were text-based adventures, I suggested we make a text-based game as the study guide. My other group members had no coding experience so I did all of the coding myself.